AbcVector: Fix "NumberFormatException: Infinite or NaN" error with large numbers

+++ src/main/groovy/org/gmetrics/metric/abc/AbcVector.groovy
@@ -50,7 +50,7 @@ class AbcVector {
         "<$assignments, $branches, $conditions>"
     }
 
-    private int squared(int val) {
+    private BigInteger squared(int val) {
         return val * val
     }
